refactor: split MCP discovery tools
